backend/pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "日本語特化ローカル音声対話アシスタント - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "sqlmodel>=0.0.22",
    "structlog>=24.4.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "soundfile>=0.13.1",
    "numpy<2",
    "reazonspeech-nemo-asr",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_assistant"]

[tool.uv.sources]
reazonspeech-nemo-asr = { git = "https://github.com/reazon-research/ReazonSpeech", subdirectory = "pkg/nemo-asr" }
